+++ src/extension.ts
@@ -1,22 +1,24 @@
-import * as vscode from 'vscode'
-import * as _file from './CreateFile'
-import * as _test from './CreateTest'
-import LensProvider from './Lens/LensProvider'
-import updateNamespace from './NamespaceUpdate'
-import * as utils from './utils'
+import * as vscode from 'vscode';
+import * as _file from './CreateFile';
+import * as _test from './CreateTest';
+import updateNamespace from './NamespaceUpdate';
+import CodeAction from './Providers/CodeAction';
+import LensProvider from './Providers/LensProvider';
+
+import * as utils from './utils';
 
 export async function activate(context) {
     /* Other -------------------------------------------------------------------- */
-    await utils.NsExtensionProviderInit()
+    await utils.NsExtensionProviderInit();
 
     /* Config ------------------------------------------------------------------- */
-    utils.setConfig()
+    utils.setConfig();
 
     vscode.workspace.onDidChangeConfiguration((event) => {
         if (event.affectsConfiguration(utils.PACKAGE_NAME)) {
-            utils.setConfig()
+            utils.setConfig();
         }
-    }, null, context.subscriptions)
+    }, null, context.subscriptions);
 
     /* Commands ----------------------------------------------------------------- */
     context.subscriptions.push(
@@ -25,29 +27,36 @@ export async function activate(context) {
         vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.generate_trait`, async (folder) => await createFile(folder, 'trait')),
         vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.generate_enum`, async (folder) => await createFile(folder, 'enum')),
         vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.generate_test_for_file`, async (e) => await _test.createTest(e)),
+        vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.extract_to_function`, async () => await _file.extractToFunction()),
+        vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.open_test_file`, async (path) => await utils.openFile(path)),
 
-        vscode.commands.registerCommand(`${utils.PACKAGE_CMND_NAME}.open_test_file`, async (path) => await utils.openFile(path))
-    )
+        vscode.languages.registerCodeActionsProvider(
+            'php',
+            new CodeAction(),
+            { providedCodeActionKinds: [vscode.CodeActionKind.RefactorExtract] },
+        ),
+        vscode.workspace.onDidRenameFiles(
+            async (event: vscode.FileRenameEvent) => await updateNamespace(event),
+        ),
+    );
 
     if (utils.config.showCodeLens) {
         context.subscriptions.push(
-            vscode.languages.registerCodeLensProvider(['php'], new LensProvider())
-        )
+            vscode.languages.registerCodeLensProvider(['php'], new LensProvider()),
+        );
     }
-
-    vscode.workspace.onDidRenameFiles(async (event: vscode.FileRenameEvent) => await updateNamespace(event))
 }
 
 async function createFile(folder, type) {
     if (folder?.path) {
         try {
-            await _file.createFile(folder.path, type)
+            await _file.createFile(folder.path, type);
         } catch (error) {
-            return
+            return;
         }
     }
 
-    await _file.insertSnippet(type)
+    await _file.insertSnippet(type);
 }
 
 export function deactivate() { }